Rename cache seconds to TTL
These changes will rename the cache seconds on the Repository interface and its implementations to $ttl rather than seconds. The reason behind this is because the TTL can be more than seconds. It can be an instance of DateTimeInterface and DateInterval as well.

Overall it also makes the internals a bit clearer so a call like this:

$seconds = $this->getSeconds($seconds);

Becomes the following which makes more sense:

$seconds = $this->getSeconds($ttl);

No changes were made to the cache stores themselves because these always accept integers/seconds so it's more appropriate to leave these as is.